rm scroll handler
